requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-dotenv>=1.0.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0
asyncpg>=0.29.0
aiohttp>=3.9.0
sympy==1.13.1
numpy==1.26.4
pynacl>=1.5.0
cryptography>=41.0.0
pytest>=7.0
pytest-asyncio>=0.21.0
tabulate>=0.9
python-multipart>=0.0.5
stripe>=5.0.0
aiosmtplib>=3.0.0
httpx>=0.28.1
python-dateutil>=2.9.0.post0
pyyaml>=6.0
uvicorn[standard]
